Updates based on comments and update to config template
## config.json.template
- Added kv pairs to reflect data required by rolemgr.py.
## CONTRIBUTING.md
- Same changes to config template as above.
## rolemgr.py
- Removed RoleStatistics class
## statistics.py
- Created this file and added RoleStatistics class
- Renamed RoleStatistics to Statistics and added cog setup
