+++ config.json.template
@@ -1,8 +1,10 @@
 {
     "token": "your bot token",
     "bot-prefix": "bot command prefix",
+    "guild-id": your guild id,
     "extensions": ["your list of extensions"],
     "extension-data": {
+        "rolemgr": {"role-setup-channel-id": 0},
         "extension name": {"whatever": "arbitrary json"}
     }
 }
